final-pjt-backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.12.0
django-extensions==3.1.5
django-seed==0.2.0
djangorestframework==3.13.1
Faker==13.3.5
idna==3.3
oauthlib==3.2.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
toposort==1.7
urllib3==1.26.9
